Back End/requirements.txt
alabaster==0.7.12
alembic==1.0.8
anaconda-client==1.7.2
anaconda-navigator==1.9.6
anaconda-project==0.8.2
asgiref==2.3.2
asn1crypto==0.24.0
astroid==2.2.5
astropy==3.1
async-timeout==3.0.1
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
autopep8==1.4.3
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==3.0.2
blinker==1.4
bokeh==1.0.2
boto==2.49.0
Bottleneck==1.2.1
CacheControl==0.12.5
cachetools==3.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
clyent==1.2.2
colorama==0.4.1
comtypes==1.1.7
conda==4.6.12
conda-build==3.17.6
conda-verify==3.1.1
constantly==15.1.0
contextlib2==0.5.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.4.2
cycler==0.10.0
Cython==0.29.2
cytoolz==0.9.0.1
dask==1.0.0
datashape==0.5.4
decorator==4.3.0
defusedxml==0.5.0
distributed==1.25.1
Django==2.1.4
django-channels==0.7.0
django-cors-headers==2.4.0
django-pandas==0.5.2
django-tastypie==0.14.2
djangorestframework==3.9.1
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.10
firebase-admin==2.15.1
Flask==0.12
Flask-Admin==1.5.0
Flask-BabelEx==0.9.3
Flask-Cors==3.0.3
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==2.0.3
Flask-Principal==0.4.0
Flask-Script==2.0.5
Flask-Security==1.7.5
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
future==0.17.1
gcloud==0.17.0
gevent==1.3.7
glob2==0.6
gmail==0.6.3
google-api-core==1.7.0
google-api-python-client==1.7.7
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.13.2
google-resumable-media==0.3.2
googleapis-common-protos==1.5.6
greenlet==0.4.15
grpcio==1.18.0
h5py==2.8.0
heapdict==1.0.0
html5lib==1.0.1
httplib2==0.12.0
hyperlink==19.0.0
idna==2.8
imageio==2.4.1
imagesize==1.1.0
importlib-metadata==0.6
incremental==17.5.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4
jedi==0.13.2
Jinja2==2.9.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.3
jupyterlab-server==0.2.0
jws==0.1.3
keyring==17.0.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
libarchive-c==2.8
llvmlite==0.26.0
locket==0.2.0
lxml==4.2.5
Mako==1.0.8
MarkupSafe==1.0
marshmallow==2.13.5
marshmallow-jsonapi==0.15.0
matplotlib==3.0.2
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.2
more-itertools==4.3.0
mpmath==1.1.0
msgpack==0.5.6
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
nltk==3.4
nose==1.3.7
notebook==5.7.4
numba==0.41.0
numexpr==2.6.8
numpy==1.15.4
numpydoc==0.8.0
oauth2client==3.0.0
oauthlib==3.0.1
odo==0.5.1
ofxhome==0.3.3
olefile==0.46
openpyxl==2.5.12
packaging==18.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
partd==0.3.9
passlib==1.7.1
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.1
pep8==1.7.1
pickleshare==0.7.5
Pillow==5.3.0
pkginfo==1.4.2
pluggy==0.8.0
ply==3.11
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.6.1
psutil==5.4.8
psycopg2==2.7.6.1
py==1.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.7.3
pycryptodomex==3.7.3
pycurl==7.43.0.2
pyflakes==2.0.0
Pygments==2.3.1
PyHamcrest==1.9.0
pylint==2.3.1
pyodbc==4.0.25
pyOpenSSL==18.0.0
pyparsing==2.3.0
PySocks==1.6.8
pytest==4.0.2
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.2.0
pytest-openfiles==0.3.1
pytest-remotedata==0.3.1
python-dateutil==2.7.5
python-editor==1.0.4
python-firebase==1.2
python-jwt==2.0.1
python-mimeparse==1.6.0
pytz==2018.7
PyWavelets==1.0.1
pywin32==223
pywinpty==0.5.5
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.5.3
qtconsole==4.4.3
QtPy==1.5.2
requests==2.21.0
requests-oauthlib==1.2.0
requests-toolbelt==0.7.0
rope==0.11.0
rsa==4.0
ruamel-yaml==0.15.46
scikit-image==0.14.1
scikit-learn==0.20.1
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
snowballstemmer==1.2.1
sortedcollections==1.0.1
sortedcontainers==2.1.0
spark-parser==1.8.7
speaklater==1.3
Sphinx==1.8.2
sphinxcontrib-websupport==1.1.0
spyder==3.3.2
spyder-kernels==0.3.0
SQLAlchemy==1.1.6
statsmodels==0.9.0
sympy==1.3
tables==3.4.4
tblib==1.3.2
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
typed-ast==1.3.1
uncompyle6==3.2.5
unicodecsv==0.14.1
Unipath==1.1
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.2.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12
widgetsnbextension==3.4.2
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
workdays==1.4
wrapt==1.10.11
WTForms==2.1
xdis==3.8.9
xlrd==1.2.0
XlsxWriter==1.1.2
xlwings==0.15.1
xlwt==1.3.0
zict==0.1.3
zope.interface==4.6.0
